+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed 
+- Remove production from MetricsAccumulator methods and add cacheHits
+
 ## [2.2.0] - 2019-03-12
 
 ## [2.2.0-beta.0] - 2019-03-11